Feat/scala 2.12 fixes (#681)
*   Move to Scala 2.12
*   Move to Kafka 2.4.1 and Confluent 5.4

Deprecated:
    * Druid Sink (not scala 2.12 compatible)
    * Elastic Sink (not scala 2.12 compatible)
    * Elastic5 Sink(not scala 2.12 compatible)

*   Redis
    *   Add support for Redis Streams

*   Cassandra
    *   Add support for setting the LoadBalancer policy on the Cassandra Sink

*   ReThinkDB
    *   Use SSL connection on Rethink initialize tables is ssl set

*   FTP Source
    *   Respect "connect.ftp.max.poll.records" when reading slices

*   MQTT Source
    *   Allow lookup of avro schema files with wildcard subscriptions

+++ kafka-connect-hive-1.1/it/src/test/scala/com/landoop/streamreactor/hive/it/HiveTests.scala
@@ -2,7 +2,7 @@ package com.landoop.streamreactor.hive.it
 
 import java.sql.{Connection, DriverManager}
 import java.time.Duration
-import java.util.{Collections, Properties}
+import java.util.Properties
 import java.util.concurrent.TimeUnit
 
 import org.apache.hadoop.conf.Configuration
@@ -16,7 +16,7 @@ import org.apache.kafka.clients.consumer.{ConsumerConfig, ConsumerRecord, KafkaC
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerConfig, ProducerRecord}
 import org.apache.kafka.common.serialization.{StringDeserializer, StringSerializer}
 import org.asynchttpclient.Dsl
-import org.scalatest.Matchers
+import org.scalatest.matchers.should.Matchers
 
 import scala.collection.JavaConverters._
 import scala.util.control.NonFatal
@@ -49,7 +49,7 @@ trait HiveTests extends Matchers {
   protected def createTopic(): String = {
     val name = "no_partition_" + Math.abs(Random.nextInt)
     Try {
-      admin.createTopics(List(new NewTopic(name, 1, 1)).asJavaCollection).all().get(30, TimeUnit.SECONDS)
+      admin.createTopics(List(new NewTopic(name, 1, 1.toShort)).asJavaCollection).all().get(30, TimeUnit.SECONDS)
       name
     }.getOrElse(sys.error(s"Could not create topic $name"))
   }
